+++ bin/integration-tests/src/tests/fpi.rs
@@ -53,7 +53,7 @@ pub async fn test_fpi_execute_program(client_config: ClientConfig) -> Result<()>
                 push.{map_key}
                 # item index
                 push.0
-                exec.::miden::account::get_map_item
+                exec.::miden::active_account::get_map_item
                 swapw dropw
             end",
             map_key = Word::from(MAP_KEY)
@@ -64,7 +64,6 @@ pub async fn test_fpi_execute_program(client_config: ClientConfig) -> Result<()>
     let code = format!(
         "
         use.miden::tx
-        use.miden::account
         begin
             # push the root of the `get_fpi_item` account procedure
             push.{proc_root}
@@ -132,7 +131,7 @@ pub async fn test_nested_fpi_calls(client_config: ClientConfig) -> Result<()> {
                 push.{map_key}
                 # item index
                 push.0
-                exec.::miden::account::get_map_item
+                exec.::miden::active_account::get_map_item
                 swapw dropw
             end",
             map_key = Word::from(MAP_KEY)
@@ -250,7 +249,7 @@ async fn standard_fpi(storage_mode: AccountStorageMode, client_config: ClientCon
                 push.{map_key}
                 # item index
                 push.0
-                exec.::miden::account::get_map_item
+                exec.::miden::active_account::get_map_item
                 swapw dropw
             end",
             map_key = Word::from(MAP_KEY)
@@ -423,6 +422,10 @@ async fn deploy_foreign_account(
         )
         .await?;
     wait_for_tx(client, tx_id).await?;
+
+    // NOTE: We get the new account state here since the first transaction updates the nonce from
+    // to 1
+    let foreign_account: Account = client.get_account(foreign_account_id).await?.unwrap().into();
 
     Ok((foreign_account, proc_root))
 }