+++ crates/web-client/src/models/mod.rs
@@ -77,6 +77,7 @@ pub mod note_tag;
 pub mod note_type;
 pub mod output_note;
 pub mod output_notes;
+pub mod package;
 pub mod partial_note;
 pub mod proven_transaction;
 pub mod provers;
@@ -102,3 +103,18 @@ pub mod transaction_status;
 pub mod transaction_store_update;
 pub mod transaction_summary;
 pub mod word;
+
+declare_js_miden_arrays! {
+    (crate::models::account::Account) -> AccountArray,
+    (crate::models::account_id::AccountId) -> AccountIdArray,
+    (crate::models::foreign_account::ForeignAccount) -> ForeignAccountArray,
+    (crate::models::note_recipient::NoteRecipient) -> NoteRecipientArray,
+    (crate::models::output_note::OutputNote) -> OutputNoteArray,
+    (crate::models::storage_slot::StorageSlot) -> StorageSlotArray,
+    (crate::models::transaction_script_inputs::TransactionScriptInputPair) -> TransactionScriptInputPairArray,
+    (crate::models::felt::Felt) -> FeltArray,
+    (crate::models::output_notes::OutputNotes) -> OutputNotesArray,
+    (crate::models::transaction_request::note_and_args::NoteAndArgs) -> NoteAndArgsArray,
+    (crate::models::transaction_request::note_details_and_tag::NoteDetailsAndTag) -> NoteDetailsAndTagArray,
+    (crate::models::transaction_request::note_id_and_args::NoteIdAndArgs) -> NoteIdAndArgsArray
+}