+++ crates/web-client/src/models/transaction_request/note_details_and_tag.rs
@@ -1,6 +1,7 @@
 use miden_client::note::{NoteDetails as NativeNoteDetails, NoteTag as NativeNoteTag};
 use wasm_bindgen::prelude::*;
 
+use crate::models::miden_arrays::NoteDetailsAndTagArray;
 use crate::models::note_details::NoteDetails;
 use crate::models::note_tag::NoteTag;
 
@@ -47,33 +48,14 @@ impl From<&NoteDetailsAndTag> for (NativeNoteDetails, NativeNoteTag) {
     }
 }
 
-#[derive(Clone)]
-#[wasm_bindgen]
-pub struct NoteDetailsAndTagArray(Vec<NoteDetailsAndTag>);
-
-#[wasm_bindgen]
-impl NoteDetailsAndTagArray {
-    #[wasm_bindgen(constructor)]
-    pub fn new(
-        note_details_and_tag_array: Option<Vec<NoteDetailsAndTag>>,
-    ) -> NoteDetailsAndTagArray {
-        let note_details_and_tag_array = note_details_and_tag_array.unwrap_or_default();
-        NoteDetailsAndTagArray(note_details_and_tag_array)
-    }
-
-    pub fn push(&mut self, note_details_and_tag: &NoteDetailsAndTag) {
-        self.0.push(note_details_and_tag.clone());
-    }
-}
-
 impl From<NoteDetailsAndTagArray> for Vec<(NativeNoteDetails, NativeNoteTag)> {
     fn from(note_details_and_tag_array: NoteDetailsAndTagArray) -> Self {
-        note_details_and_tag_array.0.into_iter().map(Into::into).collect()
+        note_details_and_tag_array.__inner.into_iter().map(Into::into).collect()
     }
 }
 
 impl From<&NoteDetailsAndTagArray> for Vec<(NativeNoteDetails, NativeNoteTag)> {
     fn from(note_details_and_tag_array: &NoteDetailsAndTagArray) -> Self {
-        note_details_and_tag_array.0.iter().map(Into::into).collect()
+        note_details_and_tag_array.__inner.iter().map(Into::into).collect()
     }
 }